upgrade - small fixes
