+++ lib/genstage_example/application.ex
@@ -1,4 +1,4 @@
-defmodule GenstageExample do
+defmodule GenstageExample.Application do
   use Application
 
   alias GenstageExample.{Producer, Consumer, Repo}
@@ -24,16 +24,4 @@ defmodule GenstageExample do
     opts = [strategy: :one_for_one, name: GenstageExample.Supervisor]
     Supervisor.start_link(children, opts)
   end
-
-  def start_later(list) do
-    GenstageExample.Task.enqueue(list)
-    notify_producer()
-  end
-
-  def start_later(module, function, args) do
-    GenstageExample.Task.enqueue({module, function, args})
-    notify_producer()
-  end
-
-  defdelegate notify_producer(), to: Producer
 end
